udev/usb-keyboards: use appropriate ATTRS fields
A pair of those rules should be created for each keyboards to support.

As pointed out by grawity on irc.freenode.net, setting keyboard map with
ENV{XKBLAYOUT} and ENV{XKBVARIANT} paired with a daemon monitoring x11
events so that xmodmap can be called properly~[1] would be better, but
after a few tests, it didn't prove to work.

[1]: https://gist.github.com/grawity/5f52f8ded53c7ffe17ed102aea226dbc